@@ -3,7 +3,7 @@ import 'dart:core';
 import 'package:flutter/foundation.dart'
     show debugDefaultTargetPlatformOverride;
 import 'package:flutter/material.dart';
-import 'package:flutter_foreground_plugin/flutter_foreground_plugin.dart';
+import 'package:flutter_background/flutter_background.dart';
 import 'package:flutter_webrtc/flutter_webrtc.dart';
 
 import 'src/data_channel_sample.dart';
@@ -24,25 +24,15 @@ void main() {
 }
 
 Future<bool> startForegroundService() async {
-  await FlutterForegroundPlugin.setServiceMethodInterval(seconds: 5);
-  await FlutterForegroundPlugin.setServiceMethod(globalForegroundService);
-  await FlutterForegroundPlugin.startForegroundService(
-    holdWakeLock: false,
-    onStarted: () {
-      print('Foreground on Started');
-    },
-    onStopped: () {
-      print('Foreground on Stopped');
-    },
-    title: 'Tcamera',
-    content: 'Tcamera sharing your screen.',
-    iconName: 'ic_stat_mobile_screen_share',
+  final androidConfig = FlutterBackgroundAndroidConfig(
+    notificationTitle: 'Title of the notification',
+    notificationText: 'Text of the notification',
+    notificationImportance: AndroidNotificationImportance.Default,
+    notificationIcon: AndroidResource(
+        name: 'background_icon',
+        defType: 'drawable'), // Default is ic_launcher from folder mipmap
   );
-  return true;
-}
-
-void globalForegroundService() {
-  debugPrint('current datetime is ${DateTime.now()}');
+  return FlutterBackground.initialize(androidConfig: androidConfig);
 }
 
 class MyApp extends StatefulWidget {
@@ -51,7 +41,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  List<RouteItem> items;
+  late List<RouteItem> items;
 
   @override
   void initState() {
