@@ -9,10 +9,10 @@ abstract class MediaStream {
   final String _ownerTag;
 
   /// The event type of this event handler is addtrack.
-  MediaTrackCallback onAddTrack;
+  MediaTrackCallback? onAddTrack;
 
   /// The event type of this event handler is removetrack.
-  MediaTrackCallback onRemoveTrack;
+  MediaTrackCallback? onRemoveTrack;
 
   String get id => _id;
 
@@ -23,7 +23,7 @@ abstract class MediaStream {
   /// The active attribute return true if this [MediaStream] is active and false otherwise.
   /// [MediaStream] is considered active if at least one of its [MediaStreamTracks] is not in the [MediaStreamTrack.ended] state.
   /// Once every track has ended, the stream's active property becomes false.
-  bool get active;
+  bool? get active;
 
   @deprecated
   Future<void> getMediaTracks();
@@ -46,12 +46,14 @@ abstract class MediaStream {
   /// The list represents a snapshot of all the [MediaStreamTrack]  objects in this stream's track set whose kind is equal to 'video'.
   List<MediaStreamTrack> getVideoTracks();
 
-  /// Returns either a [MediaStreamTrack] object from this stream's track set whose id is equal to trackId, or null, if no such track exists.
-  MediaStreamTrack getTrackById(String trackId) {
-    return getTracks().firstWhere(
-      (element) => element.id == trackId,
-      orElse: () => null,
-    );
+  /// Returns either a [MediaStreamTrack] object from this stream's track set whose id is equal to trackId, or [StateError], if no such track exists.
+  MediaStreamTrack? getTrackById(String trackId) {
+    for (var item in getTracks()) {
+      if (item.id == trackId) {
+        return item;
+      }
+    }
+    return null;
   }
 
   /// Clones the given [MediaStream] and all its tracks.
