@@ -4,8 +4,6 @@ import 'dart:js' as js;
 import 'dart:js_util' as jsutil;
 import 'dart:ui';
 
-import 'package:platform_detect/platform_detect.dart';
-
 import '../interface/media_recorder.dart';
 import '../interface/media_stream.dart';
 import '../interface/media_stream_track.dart';
@@ -13,17 +11,17 @@ import 'media_stream_impl.dart';
 import 'media_stream_track_impl.dart';
 
 class MediaRecorderWeb extends MediaRecorder {
-  html.MediaRecorder _recorder;
-  Completer<String> _completer;
-  html.CanvasElement _canvas;
-  MediaStreamTrackWeb _videoTrack;
+  html.MediaRecorder? _recorder;
+  Completer<String>? _completer;
+  late html.CanvasElement _canvas;
+  MediaStreamTrackWeb? _videoTrack;
   bool _started = false;
-  html.VideoElement _videoElement;
-  StreamSubscription _sub;
+  html.VideoElement? _videoElement;
+  StreamSubscription? _sub;
   bool _isImageCaptureSupported = false;
 
   Future<html.ImageBitmap> grabFrame() async {
-    final imageCapture = html.ImageCapture(_videoTrack.jsTrack);
+    final imageCapture = html.ImageCapture(_videoTrack!.jsTrack);
     final bitmap = await imageCapture.grabFrame();
     return bitmap;
   }
@@ -33,19 +31,21 @@ class MediaRecorderWeb extends MediaRecorder {
       return;
     }
     var bitmap;
-    num width;
-    num height;
+    int width;
+    int height;
     if (_videoElement != null) {
       bitmap = _videoElement;
-      width = _videoElement.videoWidth;
-      height = _videoElement.videoHeight;
+      width = _videoElement!.videoWidth;
+      height = _videoElement!.videoHeight;
     } else {
-      if (_videoTrack.jsTrack.readyState != 'live' || !(_videoTrack.jsTrack.enabled ?? false) || (_videoTrack.jsTrack.muted ?? false)) {
+      if (_videoTrack?.jsTrack.readyState != 'live' ||
+          !(_videoTrack?.jsTrack.enabled ?? false) ||
+          (_videoTrack?.jsTrack.muted ?? false)) {
         return;
       }
       final imageBitmap = await grabFrame();
-      width = imageBitmap.width;
-      height = imageBitmap.height;
+      width = imageBitmap.width!;
+      height = imageBitmap.height!;
       bitmap = imageBitmap;
     }
     _canvas.width = width;
@@ -64,15 +64,15 @@ class MediaRecorderWeb extends MediaRecorder {
   }
 
   html.MediaStream _setUpCanvas(html.MediaStream stream) {
-    _canvas = html.Element.canvas();
+    _canvas = html.Element.canvas() as html.CanvasElement;
     if (!_isImageCaptureSupported) {
       _videoElement = html.VideoElement()
         ..muted = true
         ..autoplay = true
         ..style.transform = 'rotateY(180deg)'
         ..srcObject = stream;
-      _videoElement.setAttribute('playsinline', '');
-      _videoElement.play();
+      _videoElement?.setAttribute('playsinline', '');
+      _videoElement?.play();
     }
     final captureStream = _canvas.captureStream();
     stream.getAudioTracks().forEach((track) {
@@ -83,25 +83,25 @@ class MediaRecorderWeb extends MediaRecorder {
 
   @override
   Future<void> start(String path,
-      {MediaStreamTrack videoTrack,
-      MediaStreamTrack audioTrack,
+      {MediaStreamTrack? videoTrack,
+      MediaStreamTrack? audioTrack,
       bool audioOnly = false,
-      int rotation,
-      Size videoSize}) {
+      int? rotation,
+      Size? videoSize}) {
     throw 'Use startWeb on Flutter Web!';
   }
 
   @override
   void startWeb(
     MediaStream stream, {
-    Function(dynamic blob, bool isLastOne) onDataChunk,
-    String mimeType = 'video/webm',
+    Function(dynamic blob, bool isLastOne)? onDataChunk,
+    String? mimeType,
     bool mirror = true,
   }) {
     var _native = stream as MediaStreamWeb;
     var videoTracks = _native.getVideoTracks();
-    if (videoTracks != null && videoTracks.isNotEmpty) {
-      _videoTrack = videoTracks.first;
+    if (videoTracks.isNotEmpty) {
+      _videoTrack = videoTracks.first as MediaStreamTrackWeb;
     }
     var mediaStream = _native.jsStream;
     _isImageCaptureSupported = checkIsImageCaptureSupported();
@@ -115,50 +115,48 @@ class MediaRecorderWeb extends MediaRecorder {
     if (onDataChunk == null) {
       var _chunks = <html.Blob>[];
       _completer = Completer<String>();
-      _recorder.addEventListener('dataavailable', (html.Event event) {
+      _recorder?.addEventListener('dataavailable', (html.Event event) {
         final html.Blob blob = js.JsObject.fromBrowserObject(event)['data'];
         if (blob.size > 0) {
           _chunks.add(blob);
         }
-        if (_recorder.state == 'inactive') {
+        if (_recorder?.state == 'inactive') {
           final blob = html.Blob(_chunks, mimeType);
           _completer?.complete(html.Url.createObjectUrlFromBlob(blob));
           _completer = null;
           _started = false;
         }
       });
-      _recorder.onError.listen((error) {
+      _recorder?.onError.listen((error) {
         _completer?.completeError(error);
         _completer = null;
         _started = false;
       });
     } else {
-      _recorder.addEventListener('dataavailable', (html.Event event) {
+      _recorder?.addEventListener('dataavailable', (html.Event event) {
         onDataChunk(
           js.JsObject.fromBrowserObject(event)['data'],
-          _recorder.state == 'inactive',
+          _recorder?.state == 'inactive',
         );
       });
     }
     _started = true;
     if (_videoElement != null) {
-      _sub = _videoElement.onPlaying.listen(
+      _sub = _videoElement?.onPlaying.listen(
         (dynamic _) {
           html.window.requestAnimationFrame(draw);
-          _recorder.start();
+          _recorder?.start();
         },
       );
     } else {
-      if (_canvas != null) {
-        html.window.requestAnimationFrame(draw);
-      }
-      _recorder.start();
+      html.window.requestAnimationFrame(draw);
+      _recorder?.start();
     }
   }
 
   @override
   Future<dynamic> stop() {
-     _started = false;
+    _started = false;
     _recorder?.stop();
     _sub?.cancel();
     _videoElement?.removeAttribute('src');
@@ -167,13 +165,9 @@ class MediaRecorderWeb extends MediaRecorder {
     return _completer?.future ?? Future.value();
   }
 
-  static bool isMirrorSupported() {
-    return !(browser.isSafari || browser.isWKWebView);
-  }
-
   bool checkIsImageCaptureSupported() {
     try {
-      html.ImageCapture(_videoTrack.jsTrack);
+      html.ImageCapture(_videoTrack!.jsTrack);
       return true;
     } catch (error) {
       //print('error:ImageCapture: $error');
