@@ -38,15 +38,16 @@ class RTCVideoRendererWeb extends VideoRenderer {
 
   static int _textureCounter = 1;
   final int _textureId;
-  html.VideoElement _videoElement;
-  MediaStream _srcObject;
+  html.VideoElement? _videoElement;
+  MediaStream? _srcObject;
   final _subscriptions = <StreamSubscription>[];
-  RTCVideoAppLifeCycleObserver _lifeCycleObserver;
+  RTCVideoAppLifeCycleObserver? _lifeCycleObserver;
 
-  set objectFit(String fit) => _videoElement.style.objectFit = fit;
+  set objectFit(String fit) => _videoElement?.style.objectFit = fit;
+  bool get mirror => value.isMirrored;
 
   set mirror(bool mirror) {
-      _videoElement.style.transform = 'rotateY(${mirror ? "180" : "0"}deg)';
+    _videoElement?.style.transform = 'rotateY(${mirror ? "180" : "0"}deg)';
   }
 
   @override
@@ -59,7 +60,7 @@ class RTCVideoRendererWeb extends VideoRenderer {
   int get textureId => _textureId;
 
   @override
-  bool get muted => value.mute ?? true;
+  bool get muted => value.mute;
 
   @override
   set muted(bool mute) {
@@ -85,7 +86,7 @@ class RTCVideoRendererWeb extends VideoRenderer {
     }
     if (_videoElement != null) {
       value = value.copyWith(isBlurred: blur);
-      _videoElement.style.filter = 'blur(${blur ? "40" : "0"}px)';
+      _videoElement?.style.filter = 'blur(${blur ? "40" : "0"}px)';
     }
   }
 
@@ -100,16 +101,16 @@ class RTCVideoRendererWeb extends VideoRenderer {
       ..style.border = 'none';
     _lifeCycleObserver?.dispose();
     _lifeCycleObserver = RTCVideoAppLifeCycleObserver(this);
-    _lifeCycleObserver.initialize();
+    _lifeCycleObserver?.initialize();
     // Allows Safari iOS to play the video inline
-    _videoElement.setAttribute('playsinline', 'true');
+    _videoElement?.setAttribute('playsinline', 'true');
 
     // ignore: undefined_prefixed_name
     ui.platformViewRegistry.registerViewFactory(
-        'RTCVideoRenderer-$textureId', (int viewId) => _videoElement);
+        'RTCVideoRenderer-$textureId', (int viewId) => _videoElement!);
 
     _subscriptions.add(
-      _videoElement.onCanPlay.listen(
+      _videoElement!.onCanPlay.listen(
         (dynamic _) {
           _updateAllValues();
           //print('RTCVideoRenderer: videoElement.onCanPlay ${value.toString()}');
@@ -118,7 +119,7 @@ class RTCVideoRendererWeb extends VideoRenderer {
     );
 
     _subscriptions.add(
-      _videoElement.onResize.listen(
+      _videoElement!.onResize.listen(
         (dynamic _) {
           _updateAllValues();
           onResize?.call();
@@ -129,17 +130,17 @@ class RTCVideoRendererWeb extends VideoRenderer {
 
     // The error event fires when some form of error occurs while attempting to load or perform the media.
     _subscriptions.add(
-      _videoElement.onError.listen(
+      _videoElement!.onError.listen(
         (html.Event _) {
           // The Event itself (_) doesn't contain info about the actual error.
           // We need to look at the HTMLMediaElement.error.
           // See: https://developer.mozilla.org/en-US/docs/Web/API/HTMLMediaElement/error
-          var error = _videoElement.error;
+          var error = _videoElement!.error;
           print('RTCVideoRenderer: videoElement.onError, ${error.toString()}');
           throw PlatformException(
-            code: _kErrorValueToErrorName[error.code],
+            code: _kErrorValueToErrorName[error] ?? '',
             message:
-                error.message != '' ? error.message : _kDefaultErrorMessage,
+                error!.message != '' ? error.message : _kDefaultErrorMessage,
             details: _kErrorValueToErrorDescription[error.code],
           );
         },
@@ -147,7 +148,7 @@ class RTCVideoRendererWeb extends VideoRenderer {
     );
 
     _subscriptions.add(
-      _videoElement.onEnded.listen(
+      _videoElement!.onEnded.listen(
         (dynamic _) {
           //print('RTCVideoRenderer: videoElement.onEnded');
         },
@@ -159,28 +160,28 @@ class RTCVideoRendererWeb extends VideoRenderer {
   void _updateAllValues() {
     value = value.copyWith(
         rotation: 0,
-        size: Size(_videoElement?.videoWidth?.toDouble() ?? 0.0,
-            _videoElement?.videoHeight?.toDouble() ?? 0.0),
+        size: Size(_videoElement?.videoWidth.toDouble() ?? 0.0,
+            _videoElement?.videoHeight.toDouble() ?? 0.0),
         renderVideo: renderVideo);
   }
 
   @override
-  MediaStream get srcObject => _srcObject;
+  MediaStream? get srcObject => _srcObject;
 
   @override
-  set srcObject(MediaStream stream) {
+  set srcObject(MediaStream? stream) {
     if (_videoElement == null) {
       throw 'Call initialize before setting the stream';
     }
     if (stream == null) {
-      _videoElement.srcObject = null;
+      _videoElement?.srcObject = null;
       _srcObject = null;
       return;
     }
     _srcObject = stream;
     var jsStream = (stream as MediaStreamWeb).jsStream;
-    _videoElement.srcObject = jsStream;
-    _videoElement.muted = stream.ownerTag == 'local';
+    _videoElement?.srcObject = jsStream;
+    _videoElement?.muted = stream.ownerTag == 'local';
     value = value.copyWith(renderVideo: renderVideo);
   }
 
@@ -205,9 +206,9 @@ class RTCVideoRendererWeb extends VideoRenderer {
   @override
   Future<bool> audioOutput(String deviceId) async {
     try {
-      if (jsutil.hasProperty(_videoElement, 'setSinkId')) {
+      if (jsutil.hasProperty(_videoElement!, 'setSinkId')) {
         await jsutil.promiseToFuture<void>(
-            jsutil.callMethod(_videoElement, 'setSinkId', [deviceId]));
+            jsutil.callMethod(_videoElement!, 'setSinkId', [deviceId]));
 
         return true;
       }
