@@ -1,4 +1,4 @@
-import 'dart:math';
+import 'dart:async';
 import 'package:flutter/material.dart';
 
 import '../interface/enums.dart';
@@ -8,58 +8,84 @@ import '../web/rtc_video_renderer_impl.dart';
 class RTCVideoView extends StatefulWidget {
   RTCVideoView(
     this._renderer, {
-    Key key,
+    Key? key,
     this.objectFit = RTCVideoViewObjectFit.RTCVideoViewObjectFitContain,
     this.mirror = false,
     this.filterQuality = FilterQuality.low,
-  })  : assert(objectFit != null),
-        assert(mirror != null),
-        assert(filterQuality != null),
-        super(key: key);
+  }) : super(key: key);
 
   final RTCVideoRenderer _renderer;
   final RTCVideoViewObjectFit objectFit;
   final bool mirror;
   final FilterQuality filterQuality;
+
   @override
   _RTCVideoViewState createState() => _RTCVideoViewState();
 }
 
 class _RTCVideoViewState extends State<RTCVideoView> {
   _RTCVideoViewState();
+
   RTCVideoRendererWeb get videoRenderer =>
       widget._renderer.delegate as RTCVideoRendererWeb;
+
   @override
   void initState() {
     super.initState();
-    widget._renderer?.delegate?.addListener(() {
-      if (mounted) setState(() {});
-    });
-  }
-
-  Widget buildVideoElementView(RTCVideoViewObjectFit objFit, bool mirror) {
-    videoRenderer.mirror = mirror;
+    widget._renderer.delegate.addListener(_onRendererListener);
+    videoRenderer.mirror = widget.mirror;
     videoRenderer.objectFit =
-        objFit == RTCVideoViewObjectFit.RTCVideoViewObjectFitContain
+        widget.objectFit == RTCVideoViewObjectFit.RTCVideoViewObjectFitContain
             ? 'contain'
             : 'cover';
+  }
+
+  void _onRendererListener() {
+    if (mounted) setState(() {});
+  }
+
+  @override
+  void dispose() {
+    widget._renderer.delegate.removeListener(_onRendererListener);
+    super.dispose();
+  }
+
+  @override
+  void didUpdateWidget(RTCVideoView oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    Timer(
+        Duration(milliseconds: 10), () => videoRenderer.mirror = widget.mirror);
+    videoRenderer.objectFit =
+        widget.objectFit == RTCVideoViewObjectFit.RTCVideoViewObjectFitContain
+            ? 'contain'
+            : 'cover';
+  }
+
+  Widget buildVideoElementView() {
+    videoRenderer.mirror = widget.mirror;
+    videoRenderer.objectFit =
+        widget.objectFit == RTCVideoViewObjectFit.RTCVideoViewObjectFitContain
+            ? 'contain'
+            : 'cover';
+    print('object FIT: ${widget.objectFit}');
     return HtmlElementView(
         viewType: 'RTCVideoRenderer-${videoRenderer.textureId}');
   }
 
-
   @override
   Widget build(BuildContext context) {
     return LayoutBuilder(
-        builder: (BuildContext context, BoxConstraints constraints) {
-      return Center(
+      builder: (BuildContext context, BoxConstraints constraints) {
+        return Center(
           child: Container(
-        width: constraints.maxWidth,
-        height: constraints.maxHeight,
-        child: widget._renderer.renderVideo
-            ? buildVideoElementView(widget.objectFit, widget.mirror)
-            : Container(),
-      ));
-    });
+            width: constraints.maxWidth,
+            height: constraints.maxHeight,
+            child: widget._renderer.renderVideo
+                ? buildVideoElementView()
+                : Container(),
+          ),
+        );
+      },
+    );
   }
 }
