@@ -7,13 +7,13 @@ abstract class MediaRecorder {
   /// For Android use audioChannel param
   /// For iOS use audioTrack
   Future<void> start(String path,
-      {MediaStreamTrack videoTrack, bool audioOnly = false, Size videoSize});
+      {MediaStreamTrack? videoTrack, bool audioOnly = false, Size? videoSize});
 
   /// Only for Flutter Web
   void startWeb(
     MediaStream stream, {
-    Function(dynamic blob, bool isLastOne) onDataChunk,
-    String mimeType,
+    Function(dynamic blob, bool isLastOne)? onDataChunk,
+    String? mimeType,
     bool mirror = true,
   });
 
