@@ -27,7 +27,7 @@ class RTCRtpTransceiverInitNative extends RTCRtpTransceiverInit {
 
   factory RTCRtpTransceiverInitNative.fromMap(Map<dynamic, dynamic> map) {
     return RTCRtpTransceiverInitNative(
-        typeStringToRtpTransceiverDirection[map['direction']],
+        typeStringToRtpTransceiverDirection[map['direction']]!,
         (map['streams'] as List<dynamic>)
             .map((e) => MediaStreamNative.fromMap(map))
             .toList(),
@@ -37,9 +37,9 @@ class RTCRtpTransceiverInitNative extends RTCRtpTransceiverInit {
   Map<String, dynamic> toMap() {
     return {
       'direction': typeRtpTransceiverDirectionToString[direction],
-      if (streams != null) 'streamIds': streams.map((e) => e.id).toList(),
+      if (streams != null) 'streamIds': streams!.map((e) => e.id).toList(),
       if (sendEncodings != null)
-        'sendEncodings': sendEncodings.map((e) => e.toMap()).toList(),
+        'sendEncodings': sendEncodings!.map((e) => e.toMap()).toList(),
     };
   }
 
@@ -47,23 +47,29 @@ class RTCRtpTransceiverInitNative extends RTCRtpTransceiverInit {
     return {
       'direction': typeRtpTransceiverDirectionToString[init.direction],
       if (init.streams != null)
-        'streamIds': init.streams.map((e) => e.id).toList(),
+        'streamIds': init.streams!.map((e) => e.id).toList(),
       if (init.sendEncodings != null)
-        'sendEncodings': init.sendEncodings.map((e) => e.toMap()).toList(),
+        'sendEncodings': init.sendEncodings!.map((e) => e.toMap()).toList(),
     };
   }
 }
 
 class RTCRtpTransceiverNative extends RTCRtpTransceiver {
-  RTCRtpTransceiverNative(this._id, this._direction, this._mid, this._sender,
-      this._receiver, _peerConnectionId);
+  RTCRtpTransceiverNative(
+    this._id,
+    this._direction,
+    this._mid,
+    this._sender,
+    this._receiver,
+    this._peerConnectionId,
+  );
 
   factory RTCRtpTransceiverNative.fromMap(Map<dynamic, dynamic> map,
-      {String peerConnectionId}) {
+      {required String peerConnectionId}) {
     var transceiver = RTCRtpTransceiverNative(
-        map['transceiverId'],
-        typeStringToRtpTransceiverDirection[map['direction']],
-        map['mid'],
+        map['transceiverId'] ?? '',
+        typeStringToRtpTransceiverDirection[map['direction']]!,
+        map['mid'] ?? '',
         RTCRtpSenderNative.fromMap(map['sender'],
             peerConnectionId: peerConnectionId),
         RTCRtpReceiverNative.fromMap(map['receiver'],
@@ -73,7 +79,7 @@ class RTCRtpTransceiverNative extends RTCRtpTransceiver {
   }
 
   static List<RTCRtpTransceiverNative> fromMaps(List<dynamic> map,
-      {String peerConnectionId}) {
+      {required String peerConnectionId}) {
     return map
         .map((e) => RTCRtpTransceiverNative.fromMap(e,
             peerConnectionId: peerConnectionId))
@@ -83,7 +89,7 @@ class RTCRtpTransceiverNative extends RTCRtpTransceiver {
   final MethodChannel _channel = WebRTC.methodChannel();
   String _peerConnectionId;
   String _id;
-  bool _stop;
+  bool _stop = false;
   TransceiverDirection _direction;
   String _mid;
   RTCRtpSender _sender;
@@ -128,12 +134,13 @@ class RTCRtpTransceiverNative extends RTCRtpTransceiver {
   @override
   Future<TransceiverDirection> getCurrentDirection() async {
     try {
-      final response = await _channel.invokeMethod(
+      final response = await WebRTC.invokeMethod(
           'rtpTransceiverGetCurrentDirection', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
         'transceiverId': _id
       });
-      _direction = typeStringToRtpTransceiverDirection[response['result']];
+
+      _direction = typeStringToRtpTransceiverDirection[response['result']]!;
       return _direction;
     } on PlatformException catch (e) {
       throw 'Unable to RTCRtpTransceiver::getCurrentDirection: ${e.message}';
@@ -147,6 +154,8 @@ class RTCRtpTransceiverNative extends RTCRtpTransceiver {
         'peerConnectionId': _peerConnectionId,
         'transceiverId': _id
       });
+
+      _stop = true;
     } on PlatformException catch (e) {
       throw 'Unable to RTCRtpTransceiver::stop: ${e.message}';
     }
