@@ -6,9 +6,14 @@ class WebRTC {
   static MethodChannel methodChannel() => _channel;
 
   static bool get platformIsDesktop =>
-      Platform.isWindows || Platform.isLinux || Platform.isMacOS;
+      Platform.isWindows ||
+      Platform.isLinux ||
+      Platform.isMacOS ||
+      Platform.isLinux;
 
   static bool get platformIsWindows => Platform.isWindows;
+
+  static bool get platformIsLinux => Platform.isLinux;
 
   static bool get platformIsMobile => Platform.isIOS || Platform.isAndroid;
 
@@ -17,4 +22,18 @@ class WebRTC {
   static bool get platformIsAndroid => Platform.isAndroid;
 
   static bool get platformIsWeb => false;
+
+  static Future<T> invokeMethod<T, P>(String methodName,
+      [dynamic param]) async {
+    var response = await _channel.invokeMethod<T>(
+      methodName,
+      param,
+    );
+
+    if (response == null) {
+      throw Exception('Invoke method: $methodName return a null response');
+    }
+
+    return response;
+  }
 }
