@@ -7,13 +7,13 @@ import 'package:flutter/foundation.dart';
 import '../interface/media_stream_track.dart';
 
 class MediaStreamTrackWeb extends MediaStreamTrack {
-  MediaStreamTrackWeb(this.jsTrack, {bool remote})
+  MediaStreamTrackWeb(this.jsTrack, {bool? remote})
       : super(
           id: jsTrack.id,
           label: jsTrack.label,
           kind: jsTrack.kind,
-          enabled: jsTrack.enabled,
-          remote: remote,
+          enabled: jsTrack.enabled ?? false,
+          remote: remote ?? false,
           switched: false,
           switching: false,
           running: true,
@@ -34,16 +34,16 @@ class MediaStreamTrackWeb extends MediaStreamTrack {
   final html.MediaStreamTrack jsTrack;
 
   @override
-  String get id => jsTrack.id;
+  String? get id => jsTrack.id;
 
   @override
-  String get kind => jsTrack.kind;
+  String? get kind => jsTrack.kind;
 
   @override
-  bool get muted => jsTrack.muted;
+  bool? get muted => jsTrack.muted;
 
   @override
-  set enabled(bool enabled) {
+  set enabled(bool? enabled) {
     var old = value.enabled;
     try {
       jsTrack.enabled = enabled;
@@ -54,26 +54,29 @@ class MediaStreamTrackWeb extends MediaStreamTrack {
   }
 
   @override
+  bool get enabled => jsTrack.enabled ?? false;
+
+  @override
   Future<bool> switchCamera() async {
     // TODO(cloudwebrtc): ???
     return false;
   }
 
   @override
-  Future<void> adaptRes(int width, int height, {int frameRate}) async {
+  Future<void> adaptRes(int width, int height, {int? frameRate}) async {
     // TODO(cloudwebrtc): ???
   }
 
   @override
   Map<String, dynamic> getConstraints() {
-    return jsTrack.getConstraints();
+    return jsTrack.getConstraints() as Map<String, dynamic>;
   }
 
   @override
-  Future<void> applyConstraints([Map<String, dynamic> constraints]) async {
+  Future<void> applyConstraints([Map<String, dynamic>? constraints]) async {
     // TODO(wermathurin): Wait for: https://github.com/dart-lang/sdk/commit/1a861435579a37c297f3be0cf69735d5b492bc6c
     // to be merged to use jsTrack.applyConstraints() directly
-    final arg = js.jsify(constraints);
+    final arg = js.jsify(constraints ?? {});
 
     final _val = await js.promiseToFuture<void>(
         js.callMethod(jsTrack, 'applyConstraints', [arg]));
@@ -120,7 +123,7 @@ class MediaStreamTrackWeb extends MediaStreamTrack {
   }
 
   @override
-  Future<dynamic> captureFrame([String filePath]) async {
+  Future<dynamic> captureFrame() async {
     final imageCapture = html.ImageCapture(jsTrack);
     final bitmap = await imageCapture.grabFrame();
     // final html.CanvasElement canvas = html.Element.canvas();
