@@ -54,8 +54,11 @@ class RTCRtpSenderWeb extends RTCRtpSender {
   Future<bool> setParameters(RTCRtpParameters parameters) async {
     try {
       var oldParameters = jsutil.callMethod(_jsRtpSender, 'getParameters', []);
-      jsutil.setProperty(oldParameters, 'encodings',
-          jsutil.jsify(parameters.encodings.map((e) => e.toMap()).toList()));
+      jsutil.setProperty(
+          oldParameters,
+          'encodings',
+          jsutil.jsify(
+              parameters.encodings?.map((e) => e.toMap()).toList() ?? []));
       return await jsutil.promiseToFuture<bool>(
           jsutil.callMethod(_jsRtpSender, 'setParameters', [oldParameters]));
     } on PlatformException catch (e) {
@@ -76,7 +79,12 @@ class RTCRtpSenderWeb extends RTCRtpSender {
   }
 
   @override
-  MediaStreamTrack get track => MediaStreamTrackWeb(_jsRtpSender.track);
+  MediaStreamTrack? get track {
+    if (null != _jsRtpSender.track) {
+      return MediaStreamTrackWeb(_jsRtpSender.track!);
+    }
+    return null;
+  }
 
   @override
   String get senderId => jsutil.getProperty(_jsRtpSender, 'senderId');
@@ -90,4 +98,6 @@ class RTCRtpSenderWeb extends RTCRtpSender {
 
   @override
   Future<void> dispose() async {}
+
+  RtcRtpSender get jsRtpSender => _jsRtpSender;
 }
